feat(cli): Complete Phase 2 - subscription group refactoring
