+++ src/sboxmgr/i18n/en.json
@@ -50,5 +50,7 @@
   "cli.error.config_save_failed": "Failed to save configuration file.",
   "cli.error.subscription_export_failed": "Failed to export subscription.",
   "cli.deprecated.list_servers": "⚠️  Warning: 'list-servers' is deprecated, use 'subscription list'",
-  "cli.deprecated.exclusions": "⚠️  Warning: 'exclusions' is deprecated, use 'subscription exclusions'"
+  "cli.deprecated.exclusions": "⚠️  Warning: 'exclusions' is deprecated, use 'subscription exclusions'",
+  "cli.subscription.help": "Manage subscriptions and servers",
+  "cli.subscription.exclusions.help": "Manage server exclusions"
 }
